+++ src/main/java/se/bjurr/prnfb/service/UserCheckService.java
@@ -2,7 +2,9 @@ package se.bjurr.prnfb.service;
 
 import static com.atlassian.bitbucket.permission.Permission.PROJECT_ADMIN;
 import static com.atlassian.bitbucket.permission.Permission.REPO_ADMIN;
+import static com.atlassian.bitbucket.permission.Permission.SYS_ADMIN;
 import static com.google.common.base.Strings.emptyToNull;
+import static com.google.common.base.Throwables.propagate;
 import static com.google.common.collect.Iterables.filter;
 import static se.bjurr.prnfb.settings.USER_LEVEL.ADMIN;
 import static se.bjurr.prnfb.settings.USER_LEVEL.EVERYONE;
@@ -19,25 +21,30 @@ import com.atlassian.bitbucket.project.Project;
 import com.atlassian.bitbucket.project.ProjectService;
 import com.atlassian.bitbucket.repository.Repository;
 import com.atlassian.bitbucket.repository.RepositoryService;
+import com.atlassian.bitbucket.user.SecurityService;
+import com.atlassian.bitbucket.util.Operation;
 import com.atlassian.sal.api.user.UserKey;
 import com.atlassian.sal.api.user.UserManager;
 import com.atlassian.sal.api.user.UserProfile;
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Predicate;
 
 public class UserCheckService {
  private final PermissionService permissionService;
  private final ProjectService projectService;
  private final RepositoryService repositoryService;
+ private final SecurityService securityService;
  private final SettingsService settingsService;
  private final UserManager userManager;
 
  public UserCheckService(PermissionService permissionService, UserManager userManager, SettingsService settingsService,
-   RepositoryService repositoryService, ProjectService projectService) {
+   RepositoryService repositoryService, ProjectService projectService, SecurityService securityService) {
   this.userManager = userManager;
   this.settingsService = settingsService;
   this.permissionService = permissionService;
   this.projectService = projectService;
   this.repositoryService = repositoryService;
+  this.securityService = securityService;
  }
 
  public Iterable<PrnfbButton> filterAllowed(List<PrnfbButton> buttons) {
@@ -60,10 +67,10 @@ public class UserCheckService {
   repositorySlug = emptyToNull(repositorySlug);
 
   if (projectKey != null && repositorySlug == null) {
-   Project project = this.projectService.getByKey(projectKey);
+   Project project = getProject(projectKey);
    return this.permissionService.hasProjectPermission(project, PROJECT_ADMIN);
   } else if (repositorySlug != null) {
-   Repository repository = this.repositoryService.getBySlug(projectKey, repositorySlug);
+   Repository repository = getRepo(projectKey, repositorySlug);
    return this.permissionService.hasRepositoryPermission(repository, REPO_ADMIN);
   }
   return isAdmin;
@@ -100,12 +107,44 @@ public class UserCheckService {
   return true;
  }
 
+ @VisibleForTesting
+ private Project getProject(String projectKey) {
+  try {
+   return this.securityService//
+     .withPermission(SYS_ADMIN, "Getting project")//
+     .call(new Operation<Project, Exception>() {
+      @Override
+      public Project perform() throws Exception {
+       return UserCheckService.this.projectService.getByKey(projectKey);
+      }
+     });
+  } catch (Exception e) {
+   throw propagate(e);
+  }
+ }
+
  private boolean isAdminAllowed(USER_LEVEL adminRestriction, @Nullable String projectKey,
    @Nullable String repositorySlug) {
   UserKey userKey = this.userManager.getRemoteUser().getUserKey();
   boolean isAdmin = isAdmin(userKey, projectKey, repositorySlug);
   boolean isSystemAdmin = isSystemAdmin(userKey);
   return isAdminAllowedCheck(adminRestriction, isAdmin, isSystemAdmin);
+ }
+
+ @VisibleForTesting
+ Repository getRepo(String projectKey, String repositorySlug) {
+  try {
+   return this.securityService//
+     .withPermission(SYS_ADMIN, "Getting repo")//
+     .call(new Operation<Repository, Exception>() {
+      @Override
+      public Repository perform() throws Exception {
+       return UserCheckService.this.repositoryService.getBySlug(projectKey, repositorySlug);
+      }
+     });
+  } catch (Exception e) {
+   throw propagate(e);
+  }
  }
 
  boolean isAdminAllowedCheck(USER_LEVEL userLevel, boolean isAdmin, boolean isSystemAdmin) {